+++ include/Uhrtypes/uhr_func_114.hpp
@@ -9,7 +9,7 @@
 
 class UHR_114_t : public iUhrType {
 public:
-    const uint8_t matrix[11][11] = {
+    const uint16_t matrix[11][11] = {
         // aus 10  11
 
         {0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10},
@@ -39,7 +39,7 @@ public:
     // 110 ->   0   111 -> 1   112 -> 2   113 -> 3
 #define LED_MAP(x)		(x > 109 ? x - 110 : 113 - (((x / 11) * 11) + (10 - (x % 11))
 
-    const uint8_t matrix[11][11] = {
+    const uint16_t matrix[11][11] = {
         {103, 104, 105, 106, 107, 108, 109, 110, 111, 112, 113},
         {102, 101, 100, 99, 98, 97, 96, 95, 94, 93, 92},
         {81, 82, 83, 84, 85, 86, 87, 88, 89, 90, 91},
@@ -56,7 +56,7 @@ public:
 
     //------------------------------------------------------------------------------
 
-    uint8_t min_arr[2][4] = {
+    uint16_t min_arr[2][4] = {
         // ergänzt aus "Uhr func 169"-datei
         {LED_MAP(110), LED_MAP(111), LED_MAP(112),
          LED_MAP(113)}, // LED für Minuten Anzeige Zeile
@@ -66,41 +66,41 @@ public:
 
     //------------------------------------------------------------------------------
 
-    virtual const uint8_t getMinArr(uint8_t col, uint8_t row) override {
+    virtual const uint16_t getMinArr(uint8_t col, uint8_t row) override {
         return min_arr[col][row];
     };
 
     //------------------------------------------------------------------------------
 
-    virtual const uint8_t getMatrix(uint8_t col, uint8_t row) override {
+    virtual const uint16_t getMatrix(uint8_t col, uint8_t row) override {
         return matrix[col][row];
     };
 
     //------------------------------------------------------------------------------
 
-    virtual const uint8_t getRMatrix(uint8_t index) override { return 0; };
+    virtual const uint16_t getRMatrix(uint16_t index) override { return 0; };
 
     //------------------------------------------------------------------------------
 
-    virtual const uint8_t getSMatrix(uint8_t index) override {
+    virtual const uint16_t getSMatrix(uint16_t index) override {
         return index;
     };
 
     //------------------------------------------------------------------------------
 
-    virtual const uint8_t NUM_PIXELS() override { return 114; };
+    virtual const uint16_t NUM_PIXELS() override { return 114; };
 
     //------------------------------------------------------------------------------
 
-    virtual const uint8_t NUM_SMATRIX() override { return 114; };
+    virtual const uint16_t NUM_SMATRIX() override { return 114; };
 
     //------------------------------------------------------------------------------
 
-    virtual const uint8_t ROWS_MATRIX() override { return 11; };
+    virtual const uint16_t ROWS_MATRIX() override { return 11; };
 
     //------------------------------------------------------------------------------
 
-    virtual const uint8_t NUM_RMATRIX() override { return 0; };
+    virtual const uint16_t NUM_RMATRIX() override { return 0; };
 
     //------------------------------------------------------------------------------
